requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
